p4i integration Integrate 105281 from 3 to main: Integrate 103775 from 3.2 to 3: Add a QT_CAST() macro def for SUN and IRIX + fix some compiler problems with o32. It does not yet compile on o32 (even when replacing ::qt_cast() calls), but it will when we test the script and get all the other stuff ironed out.
[git-p4: depot-paths = "//depot/qt/main/": change = 105282]

+++ src/tools/qdir.cpp
@@ -1250,7 +1250,7 @@ QString QDir::cleanDirPath( const QString &filePath )
 	while ( upLevel-- )
 	    newPath.insert( 0, QString::fromLatin1("/..") );
 	if ( !newPath.isEmpty() )
-	    newPath.remove( 0, 1 );
+	    newPath.remove( (uint)0, 1 );
 	else
 	    newPath = QString::fromLatin1(".");
     } else {
